Completed part 6a exercise 6.5: anecdotes, step3
