+++ part6/redux-anecdotes/src/App.js
@@ -1,33 +1,14 @@
 import { useSelector, useDispatch } from 'react-redux'
-import { addVote } from './reducers/anecdoteReducer'
 import NewAnecdote from './components/NewAnecdote'
+import AnecdoteList from './components/AnecdoteList'
 
 const App = () => {
-  const anecdotes = useSelector(state => state)
-  const dispatch = useDispatch()
-
-  const vote = (id) => {
-      dispatch(addVote(id))
-  }
-
-  return (
-    <div>
-      <h2>Anecdotes</h2>
-      {anecdotes.map(anecdote =>
-        <div key={anecdote.id}>
-          <div>
-            {anecdote.content}
-          </div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
+    return (
+        <div>
+            <AnecdoteList />
+            <NewAnecdote />
         </div>
-      )}
-      <h2>create new</h2>
-        <NewAnecdote />
-    </div>
-  )
+    )
 }
 
 export default App
