+++ part6/redux-anecdotes/src/components/AnecdoteList.js
@@ -1,0 +1,36 @@
+import { useSelector, useDispatch } from 'react-redux'
+import { addVote } from '../reducers/anecdoteReducer'
+
+const Anecdote = ({ anecdote, handleClick }) => {
+    return(
+        <div key={anecdote.id}>
+            <div>
+                {anecdote.content}
+            </div>
+            <div>
+                has {anecdote.votes}
+                <button onClick={handleClick}>vote</button>
+            </div>
+        </div>
+    )
+}
+
+const AnecdoteList = () => {
+    const anecdotes = useSelector(state => state)
+    const dispatch = useDispatch()
+
+    return (
+        <div>
+            <h2>Anecdotes</h2>
+            {anecdotes.sort((a, b) => b.votes - a.votes).map(anecdote =>
+                <Anecdote
+                anecdote={anecdote}
+                handleClick={() =>
+                    dispatch(addVote(anecdote.id))
+                }/>
+            )}
+        </div>
+    )
+}
+
+export default AnecdoteList
